"Invalid ISBN" might not be correct
openbd doesn't have a lot of books in the database. Out of ~1600 books
in my library, less than 30 were found. The ISBNs validated correctly
with isbnlib, and have metadata in other sources.

This diff updates the message to just say "not found" since opendb
returns `[null]` for both "thisisnotanisbn" and for a valid isbn that
is not found.